Turn Reduce counter to shorts
Summary: Reduce counter can be unsigned shorts instead of unsigned ints

Reviewers: richard, timh

Reviewed By: timh

Differential Revision: https://phabricator.sourcevertex.net/D5493

+++ lib/popops/codelets/reduceCodelets.cpp
@@ -67,7 +67,7 @@ public:
 
   /* The number of input regions (partials) for each output region. */
   /* This should sum to `partials.size()`. */
-  Vector<unsigned> numPartials;
+  Vector<unsigned short> numPartials;
 
   /* Multiplication factor. Might be unused. */
   float k;
